requirements.txt
fire==0.7.1
prettytable==3.17.0
pyzkaccess==1.1
pytest==8.0.0
python-dotenv==1.0.0
requests==2.31.0
termcolor==3.3.0
wcwidth==0.6.0
wrapt==1.17.3
